+++ tests/Commands/ControllerMakeCommandTest.php
@@ -92,4 +92,22 @@ class ControllerMakeCommandTest extends BaseTestCase
 
         $this->assertMatchesSnapshot($file);
     }
+
+    /** @test */
+    public function it_can_generate_a_controller_in_sub_namespace_in_correct_folder()
+    {
+        $this->artisan('module:make-controller', ['controller' => 'Api\\MyController', 'module' => 'Blog']);
+
+        $this->assertTrue(is_file($this->modulePath . '/Http/Controllers/Api/MyController.php'));
+    }
+
+    /** @test */
+    public function it_can_generate_a_controller_in_sub_namespace_with_correct_generated_file()
+    {
+        $this->artisan('module:make-controller', ['controller' => 'Api\\MyController', 'module' => 'Blog']);
+
+        $file = $this->finder->get($this->modulePath . '/Http/Controllers/Api/MyController.php');
+
+        $this->assertMatchesSnapshot($file);
+    }
 }